% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{as_posixct_compat}
\alias{as_posixct_compat}
\title{as.POSIXct ompatibility function for R < 3.5}
\usage{
as_posixct_compat(x, try_formats = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{x}{R object to be converted.}

\item{try_formats}{vector of date formats.}

\item{...}{additional arguments for as.POSIXct.}
}
\value{
a POSIXct object.
}
\description{
as.POSIXct ompatibility function for R < 3.5
}
\examples{
as_posixct_compat("2019-01-01")
}
