% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{auto_group_files}
\alias{auto_group_files}
\title{Agrupa multiples datasets de acuerdo al nombre del archivo.
Ejemplo: AAPL_1.csv, AAPL_2.csv, AAPL_3.csv agrupa a AAPL.csv}
\usage{
auto_group_files(files)
}
\arguments{
\item{files}{filepath de archivos}
}
\value{
tabla agrupada de datasets
}
\description{
Agrupa multiples datasets de acuerdo al nombre del archivo.
Ejemplo: AAPL_1.csv, AAPL_2.csv, AAPL_3.csv agrupa a AAPL.csv
}
\examples{
\donttest{
auto_group_files(c("AAPL_1.csv", "AAPL_2.csv", "AMZN_1.csv", "AMZN_2.csv"))
}
}
