% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_utils.R
\name{compatible_periodicity}
\alias{compatible_periodicity}
\title{Check if two periodicities data.frames are compatible}
\usage{
compatible_periodicity(first, second)
}
\arguments{
\item{first}{periodicity data.frame}

\item{second}{other periodicity data.frame}
}
\value{
TRUE if compatible, FALSE otherwise
}
\description{
Check if two periodicities data.frames are compatible
}
\examples{
require(xts)
data(sample_matrix)
x <- as.xts(sample_matrix)
a <- periodicity_df(head(x))
b <- periodicity_df(tail(x))

compatible_periodicity(a, b)
}
