% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_utils.R
\name{format_periodicity}
\alias{format_periodicity}
\title{Returns a formatted string of periodicity}
\usage{
format_periodicity(x, ...)
}
\arguments{
\item{x}{periodicity object}

\item{...}{additional arguments}
}
\value{
a character vector
}
\description{
Returns a formatted string of periodicity
}
\examples{
require(xts)
data(sample_matrix)
x <- as.xts(sample_matrix)

print(format_periodicity(periodicity(x)))
}
