% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{group_files}
\alias{group_files}
\title{Agrupa archivos sin tener en cuenta el formato del nombre de archivo.
Los archivos deben tener periodicidades iguales para poder ser mezclados.}
\usage{
group_files(files, name)
}
\arguments{
\item{files}{filepath de archivos}

\item{name}{nombre del dataset final}
}
\value{
tabla agrupada de los archivos
}
\description{
Agrupa archivos sin tener en cuenta el formato del nombre de archivo.
Los archivos deben tener periodicidades iguales para poder ser mezclados.
}
\examples{
\donttest{
group_files(c("PreciosApple2016.csv", "AAPL2017.csv", "StockApple2018.csv"), "AAPL")
}
}
