% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_utils.R
\name{intersect.POSIXct}
\alias{intersect.POSIXct}
\title{Set intersection between POSIXct}
\usage{
intersect.POSIXct(first, second)
}
\arguments{
\item{first}{a POSIXct vector}

\item{second}{another POSIXct vector}
}
\value{
the intersection between both vectors
}
\description{
Set intersection between POSIXct
}
\examples{
a <- as.POSIXct(sample(1:1000, 4), origin = "1970-01-01")
b <- as.POSIXct(sample(1:1000, 4), origin = "1970-01-01")
intersect.POSIXct(a, b)
}
