% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{merge_files_by_group}
\alias{merge_files_by_group}
\title{Mezcla archivos en cada grupo y guarda el csv final.}
\usage{
merge_files_by_group(grouped, save_dir, save_summary = FALSE,
  log = FALSE, info = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{grouped}{tabla agrupada}

\item{save_dir}{directorio donde guardar resultados}

\item{save_summary}{guardar el reporte como  csv}

\item{log}{mostrar logs}

\item{info}{mostrar mensajes de info}

\item{verbose}{mostrar todos los mensajes}

\item{...}{argumentos para merge_group}
}
\value{
el reporte de los resultados
}
\description{
Mezcla archivos en cada grupo y guarda el csv final.
}
\examples{
grouped <- auto_group_files(c("AAPL_1.csv", "AAPL_2.csv"))
merge_files_by_group(grouped, "processed")
}
