% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{merge_group}
\alias{merge_group}
\title{Mezcla y guarda archivos agrupados.}
\usage{
merge_group(files, group_name, save_dir = ".", verbose = FALSE, ...)
}
\arguments{
\item{files}{archivos a mezclar}

\item{group_name}{nombre del grupo}

\item{save_dir}{directorio donde guarda resultados}

\item{verbose}{muestra mensajes adicionales}

\item{...}{parametros adicionales para read_ohlcv}
}
\value{
tabla con el resumen de los archivos guardados
}
\description{
Mezcla y guarda archivos agrupados.
}
\examples{
\donttest{
files <- dir("data")
merge_group(files, "AAPL", "processed")
}
}
