% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OHLCMerge.R
\docType{package}
\name{OHLCMerge-package}
\alias{OHLCMerge}
\alias{OHLCMerge-package}
\title{OHLCMerge: funciones para mezlcar csv en formato OHLC(V)}
\description{
Merge multiple OHLC files into one.
}
\author{
\strong{Maintainer}: Jordan Ayala \email{jor.ayalag@gmail.com}

}
