% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_utils.R
\name{periodicity_df}
\alias{periodicity_df}
\title{Get periodicity in data.frame}
\usage{
periodicity_df(x)
}
\arguments{
\item{x}{a xts object}
}
\value{
data.frame with periodicity
}
\description{
Get periodicity in data.frame
}
\examples{
periodicity_df(AAPL)
}
