% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read_dataset}
\alias{read_dataset}
\title{Obtiene un dataset del formato csv}
\usage{
read_dataset(dataset, datadir, recursive = TRUE, verbose = FALSE)
}
\arguments{
\item{dataset}{nombre del dataset}

\item{datadir}{directorio en donde se encuentra el dataset}

\item{recursive}{buscar recursivamente el dataset dentro del \code{datadir}}

\item{verbose}{muestra mensajes de advertencia si es \code{TRUE}}
}
\value{
el xts con los datos
}
\description{
Obtiene un dataset del formato csv
}
\examples{
\donttest{
read_dataset("US30D1", datadir = "~/Downloads/datos_ohlc")
}
}
