% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read_ohlcv}
\alias{read_ohlcv}
\title{Read OHLCV csv}
\usage{
read_ohlcv(filename, date_format = c("\%Y-\%m-\%d", "\%Y.\%m.\%d"),
  time_format = "\%H:\%M", log = FALSE, info = FALSE)
}
\arguments{
\item{filename}{path to file}

\item{date_format}{date format}

\item{time_format}{time format}

\item{log}{should log which file is read?}

\item{info}{show additional info about file}
}
\value{
a xts
}
\description{
First column must be the date, and the second must be the time component (optional)
}
\examples{
read_ohlc("ohlc.csv")
}
