% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_utils.R
\name{setdiff.POSIXct}
\alias{setdiff.POSIXct}
\title{Set difference between POSIXct}
\usage{
setdiff.POSIXct(first, second)
}
\arguments{
\item{first}{a POSIXct vector}

\item{second}{another POSIXct vector}
}
\value{
the setdiff between both vectors
}
\description{
Set difference between POSIXct
}
\examples{
a <- as.POSIXct(sample(1:1000, 4), origin = "1970-01-01")
b <- as.POSIXct(sample(1:1000, 4), origin = "1970-01-01")
setdiff.POSIXct(a, b)
}
