% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_utils.R
\name{to_dataframe}
\alias{to_dataframe}
\title{Convierte un xts a un dataframe}
\usage{
to_dataframe(data, date_format = "\%Y-\%m-\%d",
  time_format = "\%H:\%M:\%S")
}
\arguments{
\item{data}{el xts con los datos OHLC}
}
\value{
un dataframe con datos OHLC y fecha de trading
}
\description{
Convierte un xts a un dataframe
}
\examples{
to_dataframe(DJI)
}
