% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{uses_try_formats}
\alias{uses_try_formats}
\title{Checks if this version of R has \code{try_formats} as a parameter}
\usage{
uses_try_formats()
}
\value{
TRUE if R >= 3.5
}
\description{
Checks if this version of R has \code{try_formats} as a parameter
}
\examples{
uses_try_formats()
}
