% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{valid_csv}
\alias{valid_csv}
\title{Validacion basica de csv
Si tiene header, debe tener dos filas
Si no tiene header, debe tener por lo menos una fila}
\usage{
valid_csv(filename)
}
\arguments{
\item{filename}{nombre de archivo}
}
\value{
TRUE si es valido, de lo contrario FALSE
}
\description{
Validacion basica de csv
Si tiene header, debe tener dos filas
Si no tiene header, debe tener por lo menos una fila
}
