% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_utils.R
\name{combine_xts}
\alias{combine_xts}
\title{Union los xts y retorna un xts con indices unicos}
\usage{
combine_xts(first, second, verbose = FALSE, update = FALSE)
}
\arguments{
\item{first}{un xts}

\item{second}{otro xts}

\item{verbose}{muestra indices duplicados}

\item{update}{actualizar con el segundo xts. Si es FALSE, se mantiene el primer xts.}
}
\value{
Ambos xts combinados
}
\description{
Union los xts y retorna un xts con indices unicos
}
\examples{
a <- index(AAPL[1:3, ])
b <- index(AAPL[3:5, ])
combine_xts(a, b)
}
