% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_utils.R
\name{compatible_periodicity}
\alias{compatible_periodicity}
\title{Check if two periodicities are compatible}
\usage{
compatible_periodicity(first, second)
}
\arguments{
\item{first}{periodicity data.frame}

\item{second}{other periodicity data.frame}
}
\value{
TRUE if compatible, FALSE otherwise
}
\description{
Check if two periodicities are compatible
}
\examples{
compatible_periodicty(a, b)
}
