% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{export_csv}
\alias{export_csv}
\title{Exporta el dataset a un archivo csv}
\usage{
export_csv(data, file)
}
\arguments{
\item{data}{el dataset}

\item{file}{path del archivo a exportar}
}
\value{
el archivo csv
}
\description{
Exporta el dataset a un archivo csv
}
\examples{
export_csv(DJI)
}
