% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read_ohlcv}
\alias{read_ohlcv}
\title{Read OHLCV csv}
\usage{
read_ohlcv(filename, orders = c("ymd HMS", "ymd HM"), exact = FALSE,
  log = FALSE, info = FALSE)
}
\arguments{
\item{filename}{path to file}

\item{orders}{a character vector of date-time formats. See \link[lubridate]{parse_date_time}.}

\item{exact}{use exact parsing. See \link[lubridate]{parse_date_time}.}

\item{log}{should log which file is read.}

\item{info}{show additional info about file.}
}
\value{
a xts
}
\description{
First column must be the date, and the second must be the time component (optional)
}
\examples{
\donttest{
read_ohlcv("ohlc.csv")
}
}
